back-end/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
motor==3.7.0
pydantic-settings==2.9.1
python-multipart==0.0.20
ultralytics==8.3.227
PyJWT==2.10.1
bcrypt==5.0.0
email-validator==2.3.0
certifi==2026.2.25
